Various Refactorings in preparation for Stream Validation
* Apply private attribute to methods and classes for parsing
* Remove SagaStepAction as a trait. SagaSteps can have any
* Actions don't have descriptions. Use the Example description
* Allow parsing of SubscribeAction
* Refactor ValidationState into a variety of traits
* OnClause documentation update
