+++ language/src/main/scala/com/reactific/riddl/language/parsing/ActionParser.scala
@@ -13,45 +13,41 @@ import fastparse.ScalaWhitespace.*
 /** ActionParser Define actions that various constructs can take for modelling
   * behavior in a message-passing system
   */
-trait ActionParser extends ReferenceParser with ExpressionParser {
+private[parsing] trait ActionParser extends ReferenceParser with ExpressionParser {
 
-  def arbitraryAction[u: P]: P[ArbitraryAction] = {
-    P(location ~ literalString ~ description).map { tpl =>
+  private def arbitraryAction[u: P]: P[ArbitraryAction] = {
+    P(location ~ literalString).map { tpl =>
       (ArbitraryAction.apply _).tupled(tpl)
     }
   }
 
-  def errorAction[u: P]: P[ErrorAction] = {
-    P(location ~ Keywords.error ~ literalString ~ description).map { tpl =>
+  private def errorAction[u: P]: P[ErrorAction] = {
+    P(location ~ Keywords.error ~ literalString).map { tpl =>
       (ErrorAction.apply _).tupled(tpl)
     }
   }
 
-  def setAction[u: P]: P[SetAction] = {
+  private def setAction[u: P]: P[SetAction] = {
     P(
-      Keywords.set ~/ location ~ pathIdentifier ~ Readability.to ~ expression ~
-        description
+      Keywords.set ~/ location ~ pathIdentifier ~ Readability.to ~ expression
     ).map { t => (SetAction.apply _).tupled(t) }
   }
 
-  def appendAction[u: P]: P[AppendAction] = {
+  private def appendAction[u: P]: P[AppendAction] = {
     P(
-      location ~ Keywords.append ~/ expression ~ Readability.to ~
-        pathIdentifier ~ description
+      location ~ Keywords.append ~/ expression ~ Readability.to ~ pathIdentifier
     ).map { t => (AppendAction.apply _).tupled(t) }
   }
 
-  def morphAction[u: P]: P[MorphAction] = {
+  private def morphAction[u: P]: P[MorphAction] = {
     P(
-      Keywords.morph ~/ location ~ entityRef ~ Readability.to.? ~ stateRef ~
-        description
+      Keywords.morph ~/ location ~ entityRef ~ Readability.to.? ~ stateRef
     ).map { tpl => (MorphAction.apply _).tupled(tpl) }
   }
 
-  def becomeAction[u: P]: P[BecomeAction] = {
+  private def becomeAction[u: P]: P[BecomeAction] = {
     P(
-      Keywords.become ~/ location ~ entityRef ~ Readability.to ~ handlerRef ~
-        description
+      Keywords.become ~/ location ~ entityRef ~ Readability.to ~ handlerRef
     ).map { tpl => (BecomeAction.apply _).tupled(tpl) }
   }
 
@@ -64,65 +60,63 @@ trait ActionParser extends ReferenceParser with ExpressionParser {
     }
   }
 
-  def returnAction[u: P]: P[ReturnAction] = {
-    P(Keywords.return_ ~/ location ~ expression ~ description)
+  private def returnAction[u: P]: P[ReturnAction] = {
+    P(Keywords.return_ ~/ location ~ expression )
       .map(t => (ReturnAction.apply _).tupled(t))
   }
 
-  def yieldAction[u: P]: P[YieldAction] = {
-    P(Keywords.yield_ ~/ location ~ messageConstructor ~ description)
+  private def yieldAction[u: P]: P[YieldAction] = {
+    P(Keywords.yield_ ~/ location ~ messageConstructor)
       .map(t => (YieldAction.apply _).tupled(t))
   }
 
-  def publishAction[u: P]: P[PublishAction] = {
+  private def publishAction[u: P]: P[PublishAction] = {
     P(
-      Keywords.publish ~/ location ~ messageConstructor ~ Readability.to ~
-        pipeRef ~ description
+      Keywords.publish ~/ location ~ messageConstructor ~ Readability.to ~ pipeRef
     ).map { t => (PublishAction.apply _).tupled(t) }
   }
 
-  def functionCallAction[u: P]: P[FunctionCallAction] = {
-    P(Keywords.call ~/ location ~ pathIdentifier ~ argList ~ description)
+  private def subscribeAction[u: P]: P[Action] = {
+    P(Keywords.subscribe ~/ location ~ Readability.to ~ pipeRef ~ Readability.for_ ~ typeRef).map {
+      t => (SubscribeAction.apply _).tupled(t)
+    }
+  }
+
+  private def functionCallAction[u: P]: P[FunctionCallAction] = {
+    P(Keywords.call ~/ location ~ pathIdentifier ~ argList)
       .map(tpl => (FunctionCallAction.apply _).tupled(tpl))
   }
 
-  def tellAction[u: P]: P[TellAction] = {
+  private def tellAction[u: P]: P[TellAction] = {
     P(
-      Keywords.tell ~/ location ~ messageConstructor ~ Readability.to.? ~/
-        messageTakingRef ~ description
+      Keywords.tell ~/ location ~ messageConstructor ~ Readability.to.? ~/ messageTakingRef
     ).map { t => (TellAction.apply _).tupled(t) }
   }
 
-  def replyAction[u: P]: P[ReplyAction] = {
+  private def replyAction[u: P]: P[ReplyAction] = {
     P(
-      Keywords.reply ~/ Readability.with_.? ~ location ~ messageConstructor ~
-        description
+      Keywords.reply ~/ Readability.with_.? ~ location ~ messageConstructor
     ).map { t => (ReplyAction.apply _).tupled(t) }
   }
 
-  def askAction[u: P]: P[AskAction] = {
+  private def askAction[u: P]: P[AskAction] = {
     P(
-      Keywords.ask ~/ location ~ entityRef ~ Readability.to.? ~/
-        messageConstructor ~ description
+      Keywords.ask ~/ location ~ entityRef ~ Readability.to.? ~/ messageConstructor
     ).map { tpl => (AskAction.apply _).tupled(tpl) }
   }
 
-  def compoundAction[u: P]: P[CompoundAction] = {
-    P(location ~ open ~ anyAction.rep(1, ",") ~ close ~ description)
+  private def compoundAction[u: P]: P[CompoundAction] = {
+    P(location ~ open ~ anyAction.rep(1, ",") ~ close)
       .map(tpl => (CompoundAction.apply _).tupled(tpl))
   }
 
-  def sagaStepAction[u: P]: P[SagaStepAction] = {
-    P(publishAction | tellAction | askAction | functionCallAction)
-  }
 
   def anyAction[u: P]: P[Action] = {
     P(
       replyAction | setAction | appendAction | morphAction | becomeAction |
         yieldAction | returnAction | arbitraryAction | errorAction |
-        sagaStepAction | compoundAction
+        publishAction | subscribeAction | tellAction | askAction | functionCallAction | compoundAction
     )
   }
 
-  def actionList[u: P]: P[Seq[Action]] = anyAction.rep(1)
 }
