+++ language/src/main/scala/com/reactific/riddl/language/parsing/ExpressionParser.scala
@@ -15,7 +15,7 @@ import fastparse.ScalaWhitespace.*
 import scala.collection.immutable.ListMap
 
 /** Parser rules for value expressions */
-trait ExpressionParser extends CommonParser with ReferenceParser {
+private[parsing] trait ExpressionParser extends CommonParser with ReferenceParser {
 
   // //////////////////////////////////////// Conditions == Boolean Expression
 
@@ -23,24 +23,25 @@ trait ExpressionParser extends CommonParser with ReferenceParser {
     P(terminalCondition | logicalConditions | functionCallCondition)
   }
 
-  def terminalCondition[u: P]: P[Condition] = {
+  private def terminalCondition[u: P]: P[Condition] = {
     P(trueCondition | falseCondition | arbitraryCondition)
   }
 
-  def trueCondition[u: P]: P[True] = {
+  private def trueCondition[u: P]: P[True] = {
     P(location ~ IgnoreCase("true")).map(True)./
   }
 
-  def falseCondition[u: P]: P[False] = {
+  private def falseCondition[u: P]: P[False] = {
     P(location ~ IgnoreCase("false")).map(False)./
   }
 
-  def arbitraryCondition[u: P]: P[ArbitraryCondition] = {
+  private def arbitraryCondition[u: P]: P[ArbitraryCondition] = {
     P(literalString).map(ls => ArbitraryCondition(ls.loc, ls))
   }
 
-  def emptyArgList[u: P]: P[ArgList] = { undefined[u, ArgList](ArgList()) }
-  def arguments[u: P]: P[ArgList] = {
+  private def emptyArgList[u: P]: P[ArgList] = { undefined[u, ArgList](ArgList()) }
+
+  private def arguments[u: P]: P[ArgList] = {
     P(
       emptyArgList |
         ((identifier ~ Punctuation.equalsSign ~ expression)
@@ -57,54 +58,54 @@ trait ExpressionParser extends CommonParser with ReferenceParser {
     P(Punctuation.roundOpen ~/ arguments ~ Punctuation.roundClose./)
   }
 
-  def functionCallCondition[u: P]: P[FunctionCallCondition] = {
+  private def functionCallCondition[u: P]: P[FunctionCallCondition] = {
     P(location ~ pathIdentifier ~ argList)
       .map(tpl => (FunctionCallCondition.apply _).tupled(tpl))
   }
 
-  def logicalConditions[u: P]: P[Condition] = {
+  private def logicalConditions[u: P]: P[Condition] = {
     P(
       orCondition | xorCondition | andCondition | notCondition |
         comparisonCondition
     )
   }
 
-  def orCondition[u: P]: P[OrCondition] = {
+  private def orCondition[u: P]: P[OrCondition] = {
     P(
       location ~ Operators.or ~ Punctuation.roundOpen ~/
         condition.rep(2, Punctuation.comma) ~ Punctuation.roundClose./
     ).map(t => (OrCondition.apply _).tupled(t))
   }
 
-  def xorCondition[u: P]: P[XorCondition] = {
+  private def xorCondition[u: P]: P[XorCondition] = {
     P(
       location ~ Operators.xor ~ Punctuation.roundOpen ~/
         condition.rep(2, Punctuation.comma) ~ Punctuation.roundClose./
     ).map(tpl => (XorCondition.apply _).tupled(tpl))
   }
 
-  def andCondition[u: P]: P[AndCondition] = {
+  private def andCondition[u: P]: P[AndCondition] = {
     P(
       location ~ Operators.and ~ Punctuation.roundOpen ~/
         condition.rep(2, Punctuation.comma) ~ Punctuation.roundClose./
     ).map(t => (AndCondition.apply _).tupled(t))
   }
 
-  def notCondition[u: P]: P[NotCondition] = {
+  private def notCondition[u: P]: P[NotCondition] = {
     P(
       location ~ Operators.not ~ Punctuation.roundOpen ~/ condition ~
         Punctuation.roundClose./
     ).map(t => (NotCondition.apply _).tupled(t))
   }
 
-  def comparisonCondition[u: P]: P[Comparison] = {
+  private def comparisonCondition[u: P]: P[Comparison] = {
     P(
       location ~ comparator ~ Punctuation.roundOpen ~/ expression ~
         Punctuation.comma ~ expression ~ Punctuation.roundClose./
     ).map { x => (Comparison.apply _).tupled(x) }
   }
 
-  def comparator[u: P]: P[Comparator] = {
+  private def comparator[u: P]: P[Comparator] = {
     P(StringIn("<=", "!=", "==", ">=", "<", ">")).!./.map {
       case "==" => AST.eq
       case "!=" => AST.ne
@@ -124,44 +125,44 @@ trait ExpressionParser extends CommonParser with ReferenceParser {
 
   // //////////////////////////////////////// Expressions == Any Type
 
-  def arbitraryExpression[u: P]: P[ArbitraryExpression] = {
+  private def arbitraryExpression[u: P]: P[ArbitraryExpression] = {
     P(literalString).map(ls => ArbitraryExpression(ls))
   }
 
-  def undefinedExpression[u: P]: P[UndefinedOperator] = {
+  private def undefinedExpression[u: P]: P[UndefinedOperator] = {
     P(location ~ Punctuation.undefinedMark).map(UndefinedOperator)
   }
 
-  def valueExpression[u: P]: P[ValueOperator] = {
+  private def valueExpression[u: P]: P[ValueOperator] = {
     P(location ~ Punctuation.at ~ pathIdentifier)
       .map(tpl => (ValueOperator.apply _).tupled(tpl))
   }
 
-  def aggregateConstruction[u: P]: P[AggregateConstructionExpression] = {
+  private def aggregateConstruction[u: P]: P[AggregateConstructionExpression] = {
     P(location ~ Punctuation.exclamation ~/ pathIdentifier ~ argList)
       .map(tpl => (AggregateConstructionExpression.apply _).tupled(tpl))
   }
 
-  def entityIdValue[u: P]: P[NewEntityIdOperator] = {
+  private def entityIdValue[u: P]: P[NewEntityIdOperator] = {
     P(
       location ~ Keywords.new_ ~/ Predefined.Id ~ Punctuation.roundOpen ~
         pathIdentifier ~ Punctuation.roundClose
     ).map(tpl => (NewEntityIdOperator.apply _).tupled(tpl))
   }
 
-  def terminalExpression[u: P]: P[Expression] = {
+  private def terminalExpression[u: P]: P[Expression] = {
     P(
       terminalCondition | literalDecimal | literalInteger | entityIdValue |
         valueExpression | undefinedExpression | arbitraryExpression
     )
   }
 
-  def functionCallExpression[u: P]: P[FunctionCallExpression] = {
+  private def functionCallExpression[u: P]: P[FunctionCallExpression] = {
     P(location ~ pathIdentifier ~ argList)
       .map(tpl => (FunctionCallExpression.apply _).tupled(tpl))
   }
 
-  def operatorName[u: P]: P[String] = {
+  private def operatorName[u: P]: P[String] = {
     CharPred(x => x >= 'a' && x <= 'z').! ~~ CharsWhile(x =>
       (x >= 'a' && x < 'z') ||
         (x >= 'A' && x <= 'Z') ||
@@ -169,13 +170,13 @@ trait ExpressionParser extends CommonParser with ReferenceParser {
     ).!
   }.map { case (x, y) => x + y }
 
-  def arbitraryOperator[u: P]: P[ArbitraryOperator] = {
+  private def arbitraryOperator[u: P]: P[ArbitraryOperator] = {
     P(location ~ operatorName ~ argList).map { case (loc, name, args) =>
       ArbitraryOperator(loc, LiteralString(loc, name), args)
     }
   }
 
-  def arithmeticOperator[u: P]: P[ArithmeticOperator] = {
+  private def arithmeticOperator[u: P]: P[ArithmeticOperator] = {
     P(
       location ~
         (Operators.plus.! | Operators.minus.! | Operators.times.! |
@@ -184,7 +185,7 @@ trait ExpressionParser extends CommonParser with ReferenceParser {
     ).map { tpl => (ArithmeticOperator.apply _).tupled(tpl) }
   }
 
-  def ternaryExpression[u: P]: P[Ternary] = {
+  private def ternaryExpression[u: P]: P[Ternary] = {
     P(
       location ~ Operators.if_ ~ Punctuation.roundOpen ~ condition ~
         Punctuation.comma ~ expression ~ Punctuation.comma ~ expression ~
@@ -192,14 +193,14 @@ trait ExpressionParser extends CommonParser with ReferenceParser {
     ).map(tpl => (Ternary.apply _).tupled(tpl))
   }
 
-  def groupExpression[u: P]: P[GroupExpression] = {
+  private def groupExpression[u: P]: P[GroupExpression] = {
     P(
       location ~ Punctuation.roundOpen ~/ expression.rep(0, ",") ~
         Punctuation.roundClose./
     ).map(tpl => (GroupExpression.apply _).tupled(tpl))
   }
 
-  def functionCall[u: P](
+  private def functionCall[u: P](
     names: => P[String]
   ): P[(At, String, Seq[Expression])] = {
     P(
@@ -208,25 +209,25 @@ trait ExpressionParser extends CommonParser with ReferenceParser {
     )
   }
 
-  def knownTimestamps[u: P]: P[TimeStampFunction] = {
+  private def knownTimestamps[u: P]: P[TimeStampFunction] = {
     P(functionCall(StringIn("now").!).map { tpl =>
       (TimeStampFunction.apply _).tupled(tpl)
     })
   }
 
-  def knownDates[u: P]: P[DateFunction] = {
+  private def knownDates[u: P]: P[DateFunction] = {
     functionCall(StringIn("today").!).map { tpl =>
       (DateFunction.apply _).tupled(tpl)
     }
   }
 
-  def knownNumbers[u: P]: P[NumberFunction] = {
+  private def knownNumbers[u: P]: P[NumberFunction] = {
     functionCall(StringIn("random", "pow").!).map { tpl =>
       (NumberFunction.apply _).tupled(tpl)
     }
   }
 
-  def knownStrings[u: P]: P[StringFunction] = {
+  private def knownStrings[u: P]: P[StringFunction] = {
     functionCall(StringIn("length").!).map { tpl =>
       (StringFunction.apply _).tupled(tpl)
     }
@@ -239,5 +240,4 @@ trait ExpressionParser extends CommonParser with ReferenceParser {
         knownNumbers | knownStrings | arbitraryOperator | functionCallExpression
     )
   }
-
 }
