+++ language/src/main/scala/com/reactific/riddl/language/parsing/NoWhiteSpaceParsers.scala
@@ -11,7 +11,7 @@ import fastparse.*
 import fastparse.NoWhitespace.*
 
 /** Parser rules that should not collect white space */
-trait NoWhiteSpaceParsers extends ParsingContext with Terminals {
+private[parsing] trait NoWhiteSpaceParsers extends ParsingContext with Terminals {
 
   def markdownLine[u: P]: P[LiteralString] = {
     P(
@@ -21,16 +21,16 @@ trait NoWhiteSpaceParsers extends ParsingContext with Terminals {
     ).map(tpl => (LiteralString.apply _).tupled(tpl))
   }
 
-  def hexDigit[u: P]: P[Unit] = P(CharIn("0-9a-fA-F").!)
+  private def hexDigit[u: P]: P[Unit] = P(CharIn("0-9a-fA-F").!)
 
-  def unicodeEscape[u: P]: P[Unit] =
+  private def unicodeEscape[u: P]: P[Unit] =
     P("u" ~ hexDigit ~ hexDigit ~ hexDigit ~ hexDigit).!
 
   def escape[u: P]: P[Unit] = P("\\" ~ (CharIn("\"/\\\\bfnrt") | unicodeEscape))
 
-  def stringChars(c: Char): Boolean = c != '\"' && c != '\\'
+  private def stringChars(c: Char): Boolean = c != '\"' && c != '\\'
 
-  def strChars[u: P]: P[Unit] = P(CharsWhile(stringChars))
+  private def strChars[u: P]: P[Unit] = P(CharsWhile(stringChars))
 
   def literalString[u: P]: P[LiteralString] = {
     P(
