+++ language/src/main/scala/com/reactific/riddl/language/parsing/ReferenceParser.scala
@@ -11,12 +11,13 @@ import fastparse.P
 import fastparse.StringIn
 import fastparse.ScalaWhitespace.*
 
-trait ReferenceParser extends CommonParser {
+private[parsing] trait ReferenceParser extends CommonParser {
 
   def adaptorRef[u: P]: P[AdaptorRef] = {
     P(location ~ Keywords.adaptor ~ pathIdentifier)
       .map(tpl => (AdaptorRef.apply _).tupled(tpl))
   }
+
   def commandRef[u: P]: P[CommandRef] = {
     P(location ~ Keywords.command ~ pathIdentifier)
       .map(tpl => (CommandRef.apply _).tupled(tpl))
