+++ language/src/main/scala/com/reactific/riddl/language/parsing/TypeParser.scala
@@ -13,7 +13,7 @@ import fastparse.*
 import fastparse.ScalaWhitespace.*
 
 /** Parsing rules for Type definitions */
-trait TypeParser extends CommonParser {
+private[parsing] trait TypeParser extends CommonParser {
 
   private def entityReferenceType[u: P]: P[EntityReferenceTypeExpression] = {
     P(
@@ -146,7 +146,7 @@ trait TypeParser extends CommonParser {
     P(Punctuation.roundOpen ~ integer ~ Punctuation.roundClose./).?
   }
 
-  def enumerator[u: P]: P[Enumerator] = {
+  private def enumerator[u: P]: P[Enumerator] = {
     P(location ~ identifier ~ enumValue ~ briefly ~ description).map { tpl =>
       (Enumerator.apply _).tupled(tpl)
     }
@@ -160,7 +160,7 @@ trait TypeParser extends CommonParser {
     ).map(enums => (Enumeration.apply _).tupled(enums))
   }
 
-  def alternation[u: P]: P[Alternation] = {
+  private def alternation[u: P]: P[Alternation] = {
     P(
       location ~ Keywords.one ~ Readability.of.? ~/ open ~
         (Punctuation.undefinedMark.!
@@ -221,9 +221,9 @@ trait TypeParser extends CommonParser {
   }
 
   private def makeAggregateUseCaseType(
-                                loc: At,
-                                mk: AggregateUseCase,
-                                agg: Aggregation
+    loc: At,
+    mk: AggregateUseCase,
+    agg: Aggregation
   ): AggregateUseCaseTypeExpression = { AggregateUseCaseTypeExpression(loc, mk, agg.fields) }
 
   private def aggregateUseCaseTypeExpression[u: P]: P[AggregateUseCaseTypeExpression] = {
